+++ lib/go_transitions.dart
@@ -1,5 +1,9 @@
 library go_transitions;
 
-export 'src/go_transition.dart' hide GoTransitionScope;
-export 'src/with_extensions.dart';
-export 'src/transitions.dart' show GoTransitions;
+export 'src/go_transition.dart';
+export 'src/go_transition_page.dart' hide GoTransitionScope;
+export 'src/go_transition_route.dart';
+export 'src/go_transition_settings.dart';
+export 'src/go_transition_style.dart';
+export 'src/go_transitions.dart';
+export 'src/modifiers.dart';